Remove old server code and move new one.
