+++ nengo_gui/main.py
@@ -6,7 +6,7 @@ import webbrowser
 
 import nengo_gui
 import nengo_gui.gui
-from nengo_gui.backend.backend import ModelContext
+from nengo_gui.guibackend import ModelContext, GuiServerSettings
 from nengo_gui.password import gensalt, hashpw, prompt_pw
 
 
@@ -55,13 +55,13 @@ def main():
             password = hashpw(prompt_pw(), gensalt())
         else:
             password = hashpw(args.password, gensalt())
-        server_settings = nengo_gui.backend.backend.GuiServerSettings(
+        server_settings = GuiServerSettings(
             ('', 8080), args.auto_shutdown[0], password_hash=password,
             ssl_cert=args.cert[0], ssl_key=args.key[0])
         if not server_settings.use_ssl:
             raise ValueError("Password protection only allowed with SSL.")
     else:
-        server_settings = nengo_gui.backend.backend.GuiServerSettings(
+        server_settings = GuiServerSettings(
             ('localhost', 8080), args.auto_shutdown[0], ssl_cert=args.cert[0],
             ssl_key=args.key[0])
