+++ nengo_gui/tests/test_server.py
